Integrated purescript-scores
Scores are displayed, but this is only half complete.  The canvas tag is at the moment contained in index.html and not in the Pux-generated html, which means that the display is not entirely under program control. This happens because the VexTab initialisation code currently runs on startup, which is before Pux renders its html and so Vex complains that it can see no canvas tag for its output.